Organize test files to mirror src directories
