+++ packages/core/tests/helpers/setGlobalState.test.js
@@ -1,5 +1,5 @@
 import "@testing-library/jest-dom/vitest";
-import { setGlobalState } from "../index";
+import { setGlobalState } from "../../index";
 
 document.body.innerHTML = `
   <header class="header"></header>
