--- packages/core/tests/mocks/matchMedia.mock.js
@@ -1,14 +1,0 @@
-Object.defineProperty(window, "matchMedia", {
-  writable: true,
-  value: vi.fn().mockImplementation(query => ({
-    get matches() {
-      let value = query.match(/\d+/)[0];
-      return window.innerWidth > value;
-    },
-    media: query,
-    onchange: null,
-    addEventListener: vi.fn(),
-    removeEventListener: vi.fn(),
-    dispatchEvent: vi.fn(),
-  })),
-});
