+++ packages/core/tests/utilities/Breakpoint.test.js
@@ -1,5 +1,19 @@
-import "./mocks/matchMedia.mock";
-import { Breakpoint } from "../index";
+import { Breakpoint } from "../../index";
+
+Object.defineProperty(window, "matchMedia", {
+  writable: true,
+  value: vi.fn().mockImplementation(query => ({
+    get matches() {
+      let value = query.match(/\d+/)[0];
+      return window.innerWidth > value;
+    },
+    media: query,
+    onchange: null,
+    addEventListener: vi.fn(),
+    removeEventListener: vi.fn(),
+    dispatchEvent: vi.fn(),
+  })),
+});
 
 function resizeWindow(value) {
   window.innerWidth = value;
