+++ packages/core/tests/utilities/FocusTrap.test.js
@@ -1,6 +1,6 @@
 import "@testing-library/jest-dom/vitest";
 import userEvent from "@testing-library/user-event";
-import { FocusTrap } from "../index";
+import { FocusTrap } from "../../index";
 
 const user = userEvent.setup();
